Update shiplift to 0.6
I tried to keep behavior the same.

The only changes needed to compile was a bunch of renamed field names
and shiplift 0.6 being async, so putting a few .wait() throughout.
Putting the .wait()s in did require pulling in tokio as an explicit
dependency, though the same version is already transitively included as
it's a shiplift dependency.

My primary motivation for this was just to pull in a newer version of
openssl, as I was getting "Unable to detect OpenSSL version" compilation
error because the previous openssl dependency doesn't support the
openssl installed on my system.